@@ -13,10 +13,9 @@ public final class Constants {
     public static final String UNAUTHORIZED = "UNAUTHORIZED";
     public static final String FORBIDDEN = "FORBIDDEN";
     public static final String NOT_ACCEPTABLE = "NOT_ACCEPTABLE";
-    public static final String LOGIN_NAME_REGEX = "^([a-zA-Z0-9._-]{4,16}$)";
+    public static final String LOGIN_NAME_REGEX = "^([a-zA-Z0-9._-]{6,16}$)";
     public static final String PASSWORD_REGEX = "^[!-~]{8,20}$";
-    public static final String EMAIL_REGEX = "^(?=.{1,64}@)[A-Za-z0-9_-]+(\\.[A-Za-z0-9_-]+)*@"
-            + "[^-][A-Za-z0-9-]+(\\.[A-Za-z0-9-]+)*(\\.[A-Za-z]{2,})$";
+    public static final String EMAIL_REGEX = "^[a-zA-Z0-9\\+\\.\\_\\%\\-\\+]{1,256}\\@[a-zA-Z0-9][a-zA-Z0-9\\-]{0,62}(\\.[a-zA-Z0-9][a-zA-Z0-9\\-]{0,25})+$";
     private Constants(){
     }
 }
